requirements.txt
tensorflow-cpu
flask
pydot
graphviz
jieba
numpy
termcolor
tqdm
keras
h5py
flask-cors
requests
scikit-learn
torch.nn
torch
torch.optim
pydub
sounddevice
model-community